backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic[email]
email-validator
passlib
bcrypt==3.2.2
